+++ vcv/src/plugin.hpp
@@ -1,0 +1,15 @@
+/**
+ * @file plugin.hpp
+ * @brief VCV Rack plugin declarations for South Signal Lab CLOCK.
+ * @author Axel Napolitano
+ * @copyright 2026 Axel Napolitano
+ * @license PolyForm-Noncommercial-1.0.0
+ */
+#pragma once
+
+#include <rack.hpp>
+
+using namespace rack;
+
+extern Plugin* pluginInstance;
+extern Model* modelClock;